requirements.txt
re
sys
lxml.html
requests
json
rdflib
timeit